♻️ [Refactoring]: #58 Column / ProjectConfig 型を src/types/column.ts に分離
- src/types/task.ts に同居していた Column / ProjectConfig を新規 src/types/column.ts に移設
- 18 ファイルの import パスを @/types/task → @/types/column（App.tsx のみ ./types/column）に置換
- Board feature 2 ファイルの `Column as ColumnType` エイリアスは維持（同階層 Column コンポーネントとの名前衝突回避）
- Task 型は src/types/task.ts に存続、Priority 型は @/domains/priority のまま据え置き
- 純粋な内部リファクタリング（公開 API / データ形式 / 画面挙動に影響なし）
