+++ src/features/detail/components/DetailPanel/index.tsx
@@ -6,7 +6,8 @@ import { useChildTasks } from "@/features/detail/hooks/useChildTasks";
 import { useDeleteFlow } from "@/features/detail/hooks/useDeleteFlow";
 import { useDetailLabels } from "@/features/detail/hooks/useDetailLabels";
 import { useEscToClose } from "@/features/detail/hooks/useEscToClose";
-import type { Column, Task } from "@/types/task";
+import type { Column } from "@/types/column";
+import type { Task } from "@/types/task";
 import { LabelEditor } from "../LabelEditor";
 import { MarkdownBody } from "../MarkdownBody";
 import { PrioritySelect } from "../PrioritySelect";
