+++ src/features/task-form/components/TaskForm/TaskFormStatus/__tests__/TaskFormStatus.interaction.test.tsx
@@ -1,7 +1,7 @@
 import { act, createElement } from "react";
 import { createRoot } from "react-dom/client";
 import { afterEach, expect, test, vi } from "vitest";
-import type { Column } from "@/types/task";
+import type { Column } from "@/types/column";
 import { TaskFormStatus } from "..";
 
 let container: HTMLDivElement | null = null;
