+++ src/features/task-form/components/TaskForm/__tests__/TaskForm.rendering.test.tsx
@@ -1,7 +1,8 @@
 import { act, createElement } from "react";
 import { createRoot } from "react-dom/client";
 import { afterEach, expect, test, vi } from "vitest";
-import type { Column, Task } from "@/types/task";
+import type { Column } from "@/types/column";
+import type { Task } from "@/types/task";
 import { TaskForm } from "..";
 
 let container: HTMLDivElement | null = null;
