+++ src/features/task-form/components/TaskForm/index.tsx
@@ -3,7 +3,8 @@ import { Button } from "@/components/Button";
 import { useLabelsInput } from "@/features/task-form/hooks/useLabelsInput";
 import { useTaskFormFields } from "@/features/task-form/hooks/useTaskFormFields";
 import type { TaskFormValues } from "@/features/task-form/types";
-import type { Column, Task } from "@/types/task";
+import type { Column } from "@/types/column";
+import type { Task } from "@/types/task";
 import { TaskFormActions } from "./TaskFormActions";
 import { TaskFormBody } from "./TaskFormBody";
 import { TaskFormLabels } from "./TaskFormLabels";
